api-eventos/requirements.txt
flask
redis
pika
python-dotenv